+++ tests/test_binance_order_book.py
@@ -168,7 +168,7 @@ async def test_websocket_ok(realtime_dispatcher):
     async def test_main():
         async with websockets.serve(server_main, "127.0.0.1", 0) as server:
             ws_uri = "ws://{}:{}/".format(*server.sockets[0].getsockname())
-            config_overrides = {"api": {"websockets": {"base_url": ws_uri}}}
+            config_overrides = {"api": {"ws_stream": {"base_url": ws_uri}}}
             e = exchange.Exchange(realtime_dispatcher, config_overrides=config_overrides)
             e.subscribe_to_order_book_events(p, on_order_book_10)
             e.subscribe_to_order_book_events(p, on_order_book_20, depth=20)
@@ -204,7 +204,7 @@ async def test_websocket_error(realtime_dispatcher, caplog):
     async def test_main(timeout):
         async with websockets.serve(server_main, "127.0.0.1", 0) as server:
             ws_uri = "ws://{}:{}/".format(*server.sockets[0].getsockname())
-            config_overrides = {"api": {"websockets": {"base_url": ws_uri}}}
+            config_overrides = {"api": {"ws_stream": {"base_url": ws_uri}}}
             e = exchange.Exchange(realtime_dispatcher, config_overrides=config_overrides)
             e.subscribe_to_order_book_events(p, on_order_book_event)